+++ arangod/Pregel/Worker/VertexProcessor.h
@@ -123,7 +123,7 @@ struct ActorVertexProcessor {
   std::shared_ptr<VertexComputation<V, E, M>> vertexComputation;
   std::unique_ptr<AggregatorHandler> workerAggregator;
 
-  uint32_t messageBatchSize = 5000;
+  size_t messageBatchSize = 5000;
 };
 
 }  // namespace arangodb::pregel
